# internal/contracts/shared/control-plane/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Baobab-Platform Baobab Control Plane API
  version: 1.33.0
  description: >-
    The Control Plane's runtime and administrative API (ADR-BCP-022).
    Implementations live in baobab-platform/baobab-cp. Every adminOidc
    operation also requires Control Plane authority beyond its scope (the
    platform-administrator role, or tenant-administrator membership for
    tenant-scoped operations) and, for administrative commands, a registered
    Control Plane principal; the scope is necessary, never sufficient. The
    one exception is getEffectiveAuthority, which reports the caller's own
    administrative grants and so needs no authority beyond its scope.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
servers:
  - url: https://control.baobab-platform.com/v1
    description: Production placeholder; activation requires an approved domain binding.
paths:
...
  /tenants/{tenant_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/CorrelationId"
    get:
      tags: [Tenants]
      operationId: getTenant
      summary: Read a tenant
      description: A malformed tenant id is refused with INVALID_TENANT_ID (400).
      security:
        - adminOidc: [tenant:read]
      responses:
        "200":
          description: The tenant
          content:
            application/json:
              schema:
                $ref: ./tenant.schema.json#/$defs/Tenant
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
...
    TenantId:
      name: tenant_id
      in: path
      required: true
      schema:
        $ref: ./domain.schema.json#/$defs/tenantId
...
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: A caller-supplied business correlation UUID. The service mints one when absent and returns it in problem details and emitted consequences.
      schema: { type: string, format: uuid }
...
  schemas:
    ProblemDetails:
      $ref: ../../errors/v1/problem-details.schema.json
...
  responses:
...
    InternalError:
      description: The Control Plane failed; the problem is retryable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: No such resource, or one the caller may not see
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ServiceUnavailable:
      description: A dependency is unavailable; the problem is retryable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Unresolved tenant or insufficient authority or entitlement
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  securitySchemes:
    adminOidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.baobab-platform.com/.well-known/openid-configuration
      description: Human administrator identity; issuer is environment-configured by the implementation.
    workloadOidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.baobab-platform.com/.well-known/openid-configuration
      description: Short-lived workload or delegated identity token used only over the infrastructure-owned mTLS path.